refactor: change argv-parser file structure
