+++ ciphering-cli-tool/lib/argv-parser/index.js
@@ -1,0 +1,1 @@
+module.exports = require('./argv-parser.js')
